+++ vector/src/main/java/im/vector/app/features/home/room/detail/timeline/factory/MessageItemFactory.kt
@@ -990,6 +990,7 @@ class MessageItemFactory @Inject constructor(
                         richBodyRenderer(richMessageBodyRenderer)
                         htmlPostProcessors(arrayOf<EventHtmlRenderer.PostProcessor>(pillsPostProcessor))
                         richReplyHeader(richReplyHeader)
+                        urlClickCallback(callback)
                     }
                 }
     }